pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "medusa-mcp"
version = "0.1.0"
description = "Endpoint security agent for Model Context Protocol (MCP) servers"
authors = ["Joshua Bukshpan"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/joshmaster2165/medusa"
homepage = "https://github.com/joshmaster2165/medusa"
packages = [{include = "medusa", from = "src"}]
keywords = ["mcp", "security", "agent", "gateway", "endpoint", "scanner", "ai", "model-context-protocol", "owasp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed",
]

[tool.poetry.scripts]
medusa = "medusa.cli.main:cli"
medusa-agent = "medusa.cli.agent_cli:agent_cli"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
mcp = ">=1.20.0"
pydantic = ">=2.0"
click = ">=8.0"
rich = ">=13.0"
jinja2 = ">=3.0"
httpx = ">=0.27"
pyyaml = ">=6.0"
detect-secrets = ">=1.4"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
pytest-asyncio = ">=0.23"
ruff = ">=0.4"
mypy = ">=1.10"

[tool.poetry.group.docs.dependencies]
mkdocs = ">=1.6"
mkdocs-material = ">=9.5"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
